social_media/requirements.txt
anyio==4.4.0
asgiref==3.8.1
beautifulsoup4==4.13.3
certifi==2024.7.4
crispy-bootstrap4==2024.10
distlib==0.3.8
dj-database-url==2.3.0
Django==5.1.2
django-bootstrap4==25.1
django-ckeditor==6.7.2
django-ckeditor-5==0.2.17
django-cors-headers==4.6.0
django-crispy-forms==2.3
django-environ==0.12.0
django-filter==24.3
django-heroku==0.3.1
django-js-asset==3.1.2
django-perms==2.1.0
django-storages==1.14.5
django-taggit==6.1.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.5
environ==1.0
filelock==3.16.1
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
mysql-connector==2.2.9
mysql-connector-python==9.1.0
mysqlclient==2.2.6
packaging==24.2
parts==2.1.0
pillow==11.0.0
pipenv==2024.0.2
platformdirs==4.3.6
products==2.0.1
psycopg2==2.9.10
psycopg2-binary==2.9.10
PyJWT==2.9.0
pypyodbc==1.3.6
python-telegram-bot==21.4
setuptools==75.1.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.1
virtualenv==20.26.5
whitenoise==6.8.2
python-dotenv==1.0.0
Pillow==10.0.1
